Tune zfs parameters for improved IO performance
We tune three zfs parameters:

- zfs:zfs_vdev_aggregation_limit - disable
- zfs:zfs_sync_taskq_batch_pct   - reduce to 5%
- zfs:zio_taskq_batch_pct        - reduce to 5%

Like previous zfs tuning, these are about bringing zfs parameters in
line with the reality of a modern system like the gimlet, away from
defaults that were good on a system with few CPU threads and
high-latency spinning disks.

For crucible with a 4k rand write workload, we see an 18% reduction
in CPU usage and a 15% increase in throughput.